pyproject.toml
[project]
name = "site-dcaf"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Lisboa",email = "163447861+gllisboads@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-git-revision-date-localized-plugin (>=1.5.0,<2.0.0)",
    "mkdocs-minify-plugin (>=0.8.0,<0.9.0)",
    "mkdocs-material[imaging] (>=9.6.23,<10.0.0)",
    "mkdocs-gemini-chatbot (>=1.0.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
